fix(adaptive): close the transcript bound, and ease the ExecutionStep migration
Three more from review, all correct.

**The bound had a hole I put there.** The first version returned early when a
transcript held 64 entries or fewer, on the assumption that few entries meant
small ones. `TranscriptEntry`'s fields are public, so nothing forces a harness
to build them through `bounded` — four 8 MB entries took the early return
untouched. The count check is gone; every entry is re-bounded first, then the
middle is dropped until the aggregate fits, so no construction path escapes it.

**`kind` was not counted, and my own new test caught it.** Re-bounding through
`TranscriptEntry::bounded` caps `text` and not `kind`, so four entries carrying
their payload in `kind` still reached 33 MB. It is a discriminator, not a
payload, so it is capped at 128 bytes here.

**`ExecutionStep` now derives `Default`.** The field addition is still
source-breaking for a literal that names every field — that is stated in the PR
description rather than hidden — but `..Default::default()` makes the migration
one line, which is the same courtesy `AgentRunOutcome::limit_stop`/`paused`
extend. Preferred over `#[non_exhaustive]`, which forbids the literal outright
and would be worse for a type hosts build in their own tests. `StepStatus`
derives `Default` as `Success` to allow it, and the struct's docs carry a
doctest showing the pattern.

Also moves the wire tests to `wire_tests.rs`, per the repository's rule that
Rust tests live in `_tests.rs` files rather than inline in production source.
